+++ feature/xx_old/delivery/route.go
@@ -1,7 +1,9 @@
 package delivery
 
 import (
+	"commerce-app/config"
 	"commerce-app/domain"
+	"commerce-app/feature/common"
 
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
@@ -13,9 +15,9 @@ func RouteCart(e *echo.Echo, bc domain.CartHandler) {
 		AllowHeaders: []string{echo.HeaderOrigin, echo.HeaderContentType, echo.HeaderAccept, echo.HeaderAuthorization},
 		AllowMethods: []string{echo.GET, echo.POST, echo.PUT, echo.DELETE},
 	}))
-	// e.POST("/product", bc.InsertProduct(), middleware.JWTWithConfig(common.UseJWT([]byte(config.SECRET))))
-	// e.PUT("/product/:id", bc.UpdateProduct(), middleware.JWTWithConfig(common.UseJWT([]byte(config.SECRET))))
-	// e.DELETE("/product/:id", bc.DeleteProduct(), middleware.JWTWithConfig(common.UseJWT([]byte(config.SECRET))))
+	e.POST("/cart", bc.InsertCart(), middleware.JWTWithConfig(common.UseJWT([]byte(config.SECRET))))
+	e.PUT("/cart/:id", bc.UpdateCart(), middleware.JWTWithConfig(common.UseJWT([]byte(config.SECRET))))
+	e.DELETE("/cart/:id", bc.DeleteCart(), middleware.JWTWithConfig(common.UseJWT([]byte(config.SECRET))))
 	e.GET("/cart", bc.GetAllCart())
-	// e.GET("/product/:id", bc.GetProductID())
+	e.GET("/cart/:id", bc.GetCartID())
 }
