+++ feature/xx_old/middlewares/jwt.go
@@ -1,0 +1,10 @@
+package middlewares
+
+import "github.com/labstack/echo/v4/middleware"
+
+func UseJWT(secret []byte) middleware.JWTConfig {
+	return middleware.JWTConfig{
+		SigningMethod: middleware.AlgorithmHS256,
+		SigningKey:    secret,
+	}
+}
