+++ infrastruktur/database/mysql/driver.go
@@ -2,6 +2,7 @@ package mysql
 
 import (
 	"commerce-app/config"
+	cartData "commerce-app/feature/cart/data"
 	productData "commerce-app/feature/product/data"
 	userData "commerce-app/feature/user/data"
 	"fmt"
@@ -22,5 +23,5 @@ func InitDB(cfg *config.AppConfig) *gorm.DB {
 }
 
 func MigrateData(db *gorm.DB) {
-	db.AutoMigrate(userData.User{}, productData.Product{})
+	db.AutoMigrate(userData.User{}, productData.Product{}, cartData.Cart{})
 }
